IaC/cdk-aws/requirements.txt
-e .
aws_cdk.aws_ec2
aws_cdk.aws_ecs 
aws_cdk.aws_ecr 
aws_cdk.aws_iam